requirements.txt
asgiref==3.10.0
attrs==25.4.0
coverage==7.11.0
Django==5.2.7
django-silk==5.4.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gprof2dot==2025.4.14
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
psycopg==3.2.12
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
uritemplate==4.2.0
psycopg2-binary==2.9.9